+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Added storage system to Plugin API.
 - Added requested/provided features check between plugin and platform.
 - Added OnCommand() hook from Bot to Plugin which is called when a messages with a previously registered command is received.
+- Added !help command which lists all the available commands (including commands from the CustomCommandsPlugin).
 - Call-prefix for commands is now customizable via config (default is still '!').
 
 **New storage support:**